destacame_django/requirements.txt
Django==3.1.4
asgiref==3.3.1
django-cors-headers==3.6.0
django-debug-toolbar==3.2
django-filter==2.4.0
djangorestframework==3.12.2
importlib-metadata==3.3.0
Markdown==3.3.3
psycopg2==2.8.6
pytz==2020.4
